consolidate I/UploadService
